test(autoware_traffic_light_classifier): add CnnLampRecognizer adapter test and retire characterization
Mirror the CNN classifier test layout for the lamp recognizer now that the
Node-free CnnLampRecognizerCore split has landed:

- Add test_cnn_lamp_recognizer_adapter.cpp (GPU-gated, ROS-isolated) covering
  the adapter-only concerns that cannot move to the core: the images/signals
  size guard and the per-image -> per-signal-slot scatter of getTrafficSignals.
  Mirrors test_cnn_classifier_adapter.cpp.
- Retire the real-model characterization test: the ROS-free CnnLampRecognizerCore
  unit test (#13001) plus this adapter test now cover its responsibilities, and
  dropping it removes the brittle live-model output pins (AMBER/GREEN).
- Rename the core test to the canonical test_cnn_lamp_recognizer.cpp (drop the
  _core suffix), matching the CNN naming (unsuffixed = core, _adapter = adapter),
  and switch it to ament_auto_add_gtest since it is ROS-free.

Production code is unchanged.
